requirements.txt
PyMySQL==1.1.0
pyTelegramBotAPI==4.16.1
cryptography==42.0.5
python-dotenv==1.0.1